fix TW entrypoint
Use the same pattern as Rucio ASO script
