+++ src/python/TaskWorker/MasterWorker.py
@@ -23,7 +23,6 @@ from WMCore.Configuration import loadConfigurationFile
 
 #CRAB dependencies
 from RESTInteractions import CRABRest
-import HTCondorLocator
 from ServerUtilities import newX509env
 from ServerUtilities import SERVICE_INSTANCES
 from TaskWorker import __version__
@@ -47,17 +46,6 @@ MODEURL = {'cmsweb-dev': {'host': 'cmsweb-dev.cern.ch', 'instance':  'dev'},
            'test' :{'host': None, 'instance': 'preprod'},
            'private': {'host': None, 'instance':  'dev'},
           }
-
-def validateConfig(config):
-    """Verify that the input configuration contains all needed info
-
-    :arg WMCore.Configuration config: input configuration
-    :return bool, string: flag for validation result and a message."""
-    if getattr(config, 'TaskWorker', None) is None:
-        return False, "Configuration problem: Task worker section is missing. "
-    if not hasattr(config.TaskWorker, 'scheddPickerFunction'):
-        config.TaskWorker.scheddPickerFunction = HTCondorLocator.memoryBasedChoices
-    return True, 'Ok'
 
 def getRESTParams(config, logger):
     """ get REST host name and db instance from a config object
@@ -459,63 +447,3 @@ class MasterWorker(object):
             dummyFinished = self.slaves.checkFinished()
 
         self.logger.debug("Master Worker Exiting Main Cycle.")
-
-
-if __name__ == '__main__':
-    from optparse import OptionParser
-
-    usage = "usage: %prog [options] [args]"
-    parser = OptionParser(usage=usage)
-
-    parser.add_option("-d", "--logDebug",
-                      action="store_true",
-                      dest="logDebug",
-                      default=False,
-                      help="print extra messages to stdout")
-    parser.add_option("-w", "--logWarning",
-                      action="store_true",
-                      dest="logWarning",
-                      default=False,
-                      help="don't print any messages to stdout")
-    parser.add_option("-s", "--sequential",
-                      action="store_true",
-                      dest="sequential",
-                      default=False,
-                      help="run in sequential (no subprocesses) mode")
-    parser.add_option("-c", "--console",
-                      action="store_true",
-                      dest="console",
-                      default=False,
-                      help="log to console")
-
-    parser.add_option("--config",
-                      dest="config",
-                      default=None,
-                      metavar="FILE",
-                      help="configuration file path")
-
-    (options, args) = parser.parse_args()
-
-    if not options.config:
-        raise ConfigException("Configuration not found")
-
-    configuration = loadConfigurationFile(os.path.abspath(options.config))
-    status_, msg_ = validateConfig(configuration)
-    if not status_:
-        raise ConfigException(msg_)
-
-    mw = None
-    try:
-        mw = MasterWorker(configuration, logWarning=options.logWarning, logDebug=options.logDebug, sequential=options.sequential, console=options.console)
-        signal.signal(signal.SIGINT, mw.quit_)
-        signal.signal(signal.SIGTERM, mw.quit_)
-        mw.algorithm()
-    except:
-        if mw:
-            mw.logger.exception("Unexpected and fatal error. Exiting task worker")
-        #don't really wanna miss this, propagating the exception and exiting really bad
-        raise
-    finally:
-        #there can be an exception before slaves are created, e.g. in the __init__
-        if hasattr(mw, 'slaves'):
-            mw.slaves.end()
